Fix toggling multiple lights at once
When toggling multiple Skydance lights at once (for example when activating a scene with multiple
lights), it was happening that some commands were simply ignored.
This is because Skydance Wi-Fi relay is able to handle at most X commands per second (sequentially).
Adding SequentialWriterSession fixes it as all writes goes through it and it adds a simple
sleep() after each.